// src/app/core/util/gta-3d/chunk-types.ts
export enum ChunkType {
   CHUNK_NAOBJECT = 0x0,
   CHUNK_STRUCT = 0x1,
   CHUNK_STRING = 0x2,
   CHUNK_EXTENSION = 0x3,
   CHUNK_CAMERA = 0x5,
   CHUNK_TEXTURE = 0x6,
   CHUNK_MATERIAL = 0x7,
   CHUNK_MATERIALLIST = 0x8,
   CHUNK_ATOMICSECT = 0x9,
   CHUNK_PLANESECT = 0xA,
   CHUNK_WORLD = 0xB,
   CHUNK_SPLINE = 0xC,
   CHUNK_MATRIX = 0xD,
   CHUNK_FRAMELIST = 0xE,
   CHUNK_GEOMETRY = 0xF,
   CHUNK_CLUMP = 0x10,
   CHUNK_LIGHT = 0x12,
   CHUNK_UNICODESTRING = 0x13,
   CHUNK_ATOMIC = 0x14,
   CHUNK_TEXTURENATIVE = 0x15,
   CHUNK_TEXDICTIONARY = 0x16,
   CHUNK_ANIMDATABASE = 0x17,
   CHUNK_IMAGE = 0x18,
   CHUNK_SKINANIMATION = 0x19,
   CHUNK_GEOMETRYLIST = 0x1A,
   CHUNK_ANIMANIMATION = 0x1B,
   CHUNK_HANIMANIMATION = 0x1B,
   CHUNK_TEAM = 0x1C,
   CHUNK_CROWD = 0x1D,
   CHUNK_RIGHTTORENDER = 0x1F,
   CHUNK_MTEFFECTNATIVE = 0x20,
   CHUNK_MTEFFECTDICT = 0x21,
   CHUNK_TEAMDICTIONARY = 0x22,
   CHUNK_PITEXDICTIONARY = 0x23,
   CHUNK_TOC = 0x24,
   CHUNK_PRTSTDGLOBALDATA = 0x25,
   CHUNK_ALTPIPE = 0x26,
   CHUNK_PIPEDS = 0x27,
   CHUNK_PATCHMESH = 0x28,
   CHUNK_CHUNKGROUPSTART = 0x29,
   CHUNK_CHUNKGROUPEND = 0x2A,
   CHUNK_UVANIMDICT = 0x2B,
   CHUNK_COLLTREE = 0x2C,
   CHUNK_ENVIRONMENT = 0x2D,
   CHUNK_COREPLUGINIDMAX = 0x2E,
   CHUNK_MORPH = 0x105,
   CHUNK_SKYMIPMAP = 0x110,
   CHUNK_SKIN = 0x116,
   CHUNK_PARTICLES = 0x118,
   CHUNK_HANIM = 0x11E,
   CHUNK_MATERIALEFFECTS = 0x120,
   CHUNK_PDSPLG = 0x131,
   CHUNK_ADCPLG = 0x134,
   CHUNK_UVANIMPLG = 0x135,
   CHUNK_BINMESH = 0x50E,
   CHUNK_NATIVEDATA = 0x510,
   CHUNK_VERTEXFORMAT = 0x510,
   CHUNK_PIPELINESET = 0x253F2F3,
   CHUNK_SPECULARMAT = 0x253F2F6,
   CHUNK_2DFX = 0x253F2F8,
   CHUNK_NIGHTVERTEXCOLOR = 0x253F2F9,
   CHUNK_COLLISIONMODEL = 0x253F2FA,
   CHUNK_REFLECTIONMAT = 0x253F2FC,
   CHUNK_MESHEXTENSION = 0x253F2FD,
   CHUNK_FRAME = 0x253F2FE,
}